inlined all definitions
