--- types/scrivito/lib/Attribute.d.ts
@@ -1,23 +1,0 @@
-export type Attribute =
-    'date' |
-    'datetime' |
-    'enum' |
-    'multienum' |
-    'html' |
-    'string' |
-    'stringlist' |
-    'integer' |
-    'float' |
-    'binary' |
-    'link' |
-    'linklist' |
-    'reference' |
-    'referencelist' |
-    'widgetlist';
-
-export interface AttributeOptions {
-    values?: any[];
-    only?: string | string[];
-}
-
-export type AttributeWithOptions = [Attribute, AttributeOptions];
