--- types/scrivito/lib/Config.d.ts
@@ -1,12 +1,0 @@
-import { Obj } from './Obj';
-
-export interface ConfigOptions {
-    tenant: string;
-    homepage: () => Obj;
-    origin?: string;
-    routingBasePath?: string;
-    visitorAuthentication?: boolean;
-    // Hard to type
-    constraintsValidation?: (constraints: any) => any;
-    endpoint?: string;
-}
