--- types/scrivito/lib/MetadataCollection.d.ts
@@ -1,4 +1,0 @@
-export class MetadataCollection {
-    private constructor();
-    get(name: string): string | string[] | number | Date | null;
-}
